+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Change Log
 
 1.3.57 (pre-release) [2024-06-10]
 
+- Added dialogs for selection/entering the missing origin's API token.
 - Fixed missing API error response message, where incorrect `response: [object Object]` was displayed instead of actual response.
 - Fixed creation of universal module.